// CLIENT/src/redux/actionTypes/ModalActionTypes.ts
export enum EModalActionTypes {
  OPEN_CONSULT_MODAL = "OPEN_CONSULT_MODAL",
  OPEN_SHIPMENT_MODAL = "OPEN_SHIPMENT_MODAL",
  OPEN_PAYMENT_MODAL = "OPEN_PAYMENT_MODAL",
  OPEN_ORDER_MODAL = "OPEN_ORDER_MODAL",
  OPEN_REGISTER_MODAL = "OPEN_REGISTER_MODAL",
  OPEN_PASSWORD_MODAL = "OPEN_PASSWORD_MODAL",
  OPEN_PRODUCT_LIST_MODAL = "OPEN_PRODUCT_LIST_MODAL",
  OPEN_ADD_PRODUCT_MODAL = "OPEN_ADD_PRODUCT_MODAL",
  OPEN_ADD_SHIPMENT_MODAL = "OPEN_ADD_SHIPMENT_MODAL",
  OPEN_RATING_MODAL = "OPEN_RATING_MODAL",
  OPEN_TRAILER_MODAL = "OPEN_TRAILER_MODAL",
  OPEN_BOOKING_TICKET_MODAL = "OPEN_BOOKING_TICKET_MODAL",
  OPEN_TIMEOUT_MODAL = "OPEN_TIMEOUT_MODAL",
  CLOSE_CONSULT_MODAL = "CLOSE_CONSULT_MODAL",
  CLOSE_SHIPMENT_MODAL = "CLOSE_SHIPMENT_MODAL",
  CLOSE_PAYMENT_MODAL = "CLOSE_PAYMENT_MODAL",
  CLOSE_ORDER_MODAL = "CLOSE_ORDER_MODAL",
  CLOSE_REGISTER_MODAL = "CLOSE_REGISTER_MODAL",
  CLOSE_PASSWORD_MODAL = "CLOSE_PASSWORD_MODAL",
  CLOSE_PRODUCT_LIST_MODAL = "CLOSE_PRODUCT_LIST_MODAL",
  CLOSE_ADD_PRODUCT_MODAL = "CLOSE_ADD_PRODUCT_MODAL",
  CLOSE_ADD_SHIPMENT_MODAL = "CLOSE_ADD_SHIPMENT_MODAL",
  CLOSE_RATING_MODAL = "CLOSE_RATING_MODAL",
  CLOSE_TRAILER_MODAL = "CLOSE_TRAILER_MODAL",
  CLOSE_BOOKING_TICKET_MODAL = "CLOSE_BOOKING_TICKET_MODAL",
  CLOSE_TIMEOUT_MODAL = "CLOSE_TIMEOUT_MODAL",
}